server/src/requirements.txt
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pymysql
python-dotenv
alembic
psycopg2-binary
slowapi
email-validator